feat(session): v0.0.1
Implements sessions closing logic and billing resume

+++ application/core/src/main/kotlin/br/com/fairie/partypay/usecase/menu/vo/Order.kt
@@ -1,7 +1,9 @@
 package br.com.fairie.partypay.usecase.menu.vo
 
+import java.math.BigDecimal
+
 class Order(
     val name: String,
     val description: String,
-    val value: Double
+    val value: BigDecimal
 )
